feat: CoreData Entity 정의, CoreDataManager 타입 구현
