+++ Diary+CoreDataClass.swift
@@ -1,0 +1,15 @@
+//
+//  Diary+CoreDataClass.swift
+//  Diary
+//
+//  Created by Harry on 2023/04/28.
+//
+//
+
+import Foundation
+import CoreData
+
+@objc(Diary)
+public class Diary: NSManagedObject {
+
+}
