+++ Diary/CoreData/CoreDataManager.swift
@@ -1,0 +1,37 @@
+//
+//  CoreDataManager.swift
+//  Diary
+//
+//  Created by Harry on 2023/04/28.
+//
+
+import Foundation
+import CoreData
+
+final class CoreDataManager {
+    static var shared: CoreDataManager = CoreDataManager()
+
+    lazy var persistentContainer: NSPersistentContainer = {
+        let container = NSPersistentContainer(name: "Diary")
+        container.loadPersistentStores(completionHandler: { (storeDescription, error) in
+            if let error = error as NSError? {
+                fatalError("Unresolved error \(error), \(error.userInfo)")
+            }
+        })
+        return container
+    }()
+
+    func saveContext () {
+        let context = persistentContainer.viewContext
+        
+        if context.hasChanges {
+            do {
+                try context.save()
+            } catch {
+                let nserror = error as NSError
+                fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
+            }
+        }
+    }
+
+}
